google geo converter, auditing

+++ src/main/java/com/outofmemory/dto/LocationDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class LocationDto {
-    private int lng;
-    private int lat;
+    private Double lng;
+    private Double lat;
 }
